Make cookbook work with Chef 13. (#1)
* Upgrades AWS SDK to V2.

This commit upgrades the AWS SDK to version 2 and adds a test suite for testing artifact deploys with Amazon s3.

To test the artifact deploys, you'll need to create a databag for that test suite and/or deploy to Amazon. First, it will try to authorize the user through databags. If it is unable to find an aws databag, it will switch to using the instance profile to pull the informmation directly from the Amazon EC2 instance.

* Make cookbook work with Chef 13.